Generics v5.2 (static check methods)

+++ HW_17/src/expression/type/UncheckedIntegerEType.java
@@ -4,7 +4,7 @@ import expression.exceptions.*;
 
 public class UncheckedIntegerEType extends AbstractEType<Integer> implements ForbiddenDivisionByZero<Integer> {
     public UncheckedIntegerEType(Integer value) {
-        super(value.intValue());
+        super(value);
     }
 
     @Override
